+++ gem/engine/__init__.py
@@ -24,6 +24,7 @@ from datetime import timedelta
 from os import W_OK
 from os import mkdir
 from os import access
+from os import getenv
 from os import getpid
 from os import remove
 from os import makedirs